pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mangadown"
version = "0.2.0"
description = "Automated manga downloader for Kindle"
requires-python = ">=3.12"
license = "Beerware"
authors = [{ name = "Antoine Lecubin", email = "antoinelecubin@msn.com" }]
dependencies = [
    "beautifulsoup4>=4.12",
    "click>=8.1",
    "cloudscraper>=1.2.71",
    "httpx>=0.27",
    "KindleComicConverter @ git+https://github.com/ciromattia/kcc.git",
    "packaging>=20.0",
    "Pillow>=10.0",
    "platformdirs>=4.0",
    "pydantic-settings>=2.0",
    "tenacity>=8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.4",
    "ty",
]

[project.scripts]
mangadown = "mangadown.cli.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
